Module Yocaml_syndication.Category

Conveys information about a category associated with an entry or feed. This specification assigns no meaning to the content (if any) of this element.

see https://datatracker.ietf.org/doc/html/rfc4287#section-4.2.2

Types

type t

Helpers

val make : ?scheme:string -> ?label:string -> string -> t

make ?scheme ?label term constructs an element of t. scheme is used as domain for rss2 and label is discarded.

val to_atom : t -> Xml.node

Generate an Atom node.

val to_rss2 : t -> Xml.node

Generate a Rss2 node.