Module Yocaml_syndication.Datetime

An extension of Yocaml.Archetype.Datetime.t with Timezone support.

Types

type t

Helpers

val compare : t -> t -> int

compare dt1 dt2 a dummy comparison function that does not care about the Timezone (because we expect that every RSS/Atom elements are published on the same Timezone).

val make : ?tz:Tz.t -> Yocaml.Datetime.t -> t

make ?tz datetime build a datetime associated with a Timezone.

val to_string : t -> string

to_string date converts a date into a string.

val to_string_rfc3339 : t -> string

to_string_rfc3339 date converts a date into a string according to the rfc 3339.