Module Yocaml_syndication.Tz

Describes (almost) a Timezone according to RFC822.

Types

type t =

| Ut

| Gmt

| Est

| Edt

| Cst

| Cdt

| Mst

| Mdt

| Pst

| Pdt

| Plus of int

| Minus of int

A type describing a timezone.

Helpers

val plus : int -> t

plus 200 generates the TZ "+0200".

val minus : int -> t

minus 200 generates the TZ "-0200".

val to_string : t -> string

to_string tz render a string representation of a timezone.

val to_string_rfc3339 : t -> string

to_string_rfc3339 tz render a string representation of a timezone (computing fixed timezone into offset).